requirements.txt
streamlit== 1.22.0
plotly==5.14.1
pandas==1.5.3
numpy==1.24.2
folium==0.14.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.11.1
Pillow==9.4.0
inflection==0.5.1
image==1.5.33